+++ src/Page/Home.tsx
@@ -1,23 +1,11 @@
-import React, { useEffect } from 'react';
+import React  from 'react';
 import Header from "../Components/Layout/Header.js";
 import HomeComponent from "../Components/ContentMenu/HomeComponent";
 import Footer from "../Components/Layout/Footer";
-import { useAuth } from "../Components/hooks/use-auth";
-import { useNavigate } from "react-router-dom";
+
 
 const Home: React.FC = () => {
-    // const { isAuth,  } = useAuth();
-    // const navigate = useNavigate();
 
-    // useEffect(() => {
-    //     if (!isAuth) {
-    //         navigate("/login");
-    //     }
-    // }, [isAuth, navigate]);
-    //
-    // if (!isAuth) {
-    //     return null;
-    // }
 
     return (
         <>
