+++ src/fire_base.js
@@ -1,6 +1,5 @@
 import { initializeApp } from "firebase/app";
 // import { getAnalytics } from "firebase/analytics";
-// console.log(import.meta.env)
 const firebaseConfig = {
     apiKey: import.meta.env.VITE_APP_FIREBACE_API_KEY,
     authDomain: import.meta.env.VITE_APP_AUTH_DOMAIN,
